# Resort the task queue after a task has been assigned.