Bugfix! I broke Joona's algorithm :frowning_face: 

Resorting is important because the task queue is only regenerated every minute.

The resorting was accidentally deleted in the commit "dynamic_tasks".