@ppigazzini 2021-09-12
DEV updated, joined some workers, started few test with throughput 25-50-100%