@ppigazzini 2021-09-12
All tests with priority 9, throughput 25-50-100% (100% not indicated)
<html>
<body>
<!--StartFragment--><h3 style="margin: 10px 0px; font-family: &quot;Helvetica Neue&quot;, Helvetica, Arial, sans-serif; font-weight: bold; line-height: 40px; color: rgb(177, 173, 167); text-rendering: optimizelegibility; font-size: 24.5px; font-style: normal; font-variant-ligatures: normal; font-variant-caps: normal; letter-spacing: normal; orphans: 2; text-align: start; text-indent: 0px; text-transform: none; white-space: normal; widows: 2; word-spacing: 0px; -webkit-text-stroke-width: 0px; background-color: rgb(17, 17, 17); text-decoration-thickness: initial; text-decoration-style: initial; text-decoration-color: initial;">Active - 9 tests</h3>

21-09-12 | use | master | diff | LLR: 0.77 (-2.94,2.94) <-0.50,2.50>Total: 1312 W: 440 L: 375 D: 497Ptnml(0-2): 34, 130, 282, 157, 53 | sprt @ 0.5+0.005 th 1cores: 4 | #1026 SPRT - throughput 25%
-- | -- | -- | -- | -- | -- | --
21-09-12 | use | master | diff | LLR: 0.22 (-2.94,2.94) <-0.50,2.50>Total: 1136 W: 365 L: 344 D: 427Ptnml(0-2): 35, 134, 219, 135, 45 | sprt @ 0.5+0.005 th 1cores: 5 | #1026 NumGames - Throughput 25%
21-09-12 | use | master | diff | LLR: -0.06 (-2.94,2.94) <-0.50,2.50>Total: 2080 W: 652 L: 653 D: 775Ptnml(0-2): 61, 242, 437, 237, 63 | sprt @ 0.5+0.005 th 1cores: 6 | #1026 NumGames - Throughput 50%
21-09-12 | use | master | diff | LLR: -0.14 (-2.94,2.94) <-0.50,2.50>Total: 2016 W: 640 L: 647 D: 729Ptnml(0-2): 74, 211, 434, 226, 63 | sprt @ 0.5+0.005 th 1cores: 6 | #1026 SPRT - throughput 50%
21-09-12 | use | master | diff | LLR: -0.20 (-2.94,2.94) <-0.50,2.50>Total: 5048 W: 1567 L: 1573 D: 1908Ptnml(0-2): 144, 570, 1106, 556, 148 | sprt @ 0.5+0.005 th 1cores: 12 | #1026 SPRT
21-09-12 | use | master | diff | LLR: -0.80 (-2.94,2.94) <-0.50,2.50>Total: 4512 W: 1395 L: 1451 D: 1666Ptnml(0-2): 165, 512, 938, 496, 145 | sprt @ 0.5+0.005 th 1cores: 9 | #1026 NumGames
21-09-12 | use | moves1 | diff | 916/30000 iterations1832/60000 games played | 60000 @ 0.5+0.005 th 1cores: 8 | #1026 SPSA
21-09-12 | use | moves1 | diff | 414/30000 iterations828/60000 games played | 60000 @ 0.5+0.005 th 1cores: 6 | #1026 SPSA - Throughput 50%
21-09-12 | use | moves1 | diff | 272/30000 iterations544/60000 games played | 60000 @ 0.5+0.005 th 1cores: 3 | #1026 SPSA - Throughput 25%

<!--EndFragment-->
</body>
</html>