@vdbergh 2021-09-12
Thanks for testing. It is hard to conclude anything on the short term. The throughput algorithm should keep the average throughput constant (roughly speaking), but on the sort term there can be a lot of variation.

The bug I introduced is that when a run is in front of the queue for a new task, the queue is not resorted for some time, and so the same run stays in front, hence it will also be dispensing the next task, and so on. I noticed this when I saw a run suddenly grabbing a lot of cores.