@vdbergh 2021-09-13
Thanks. Unfortunately it is still happening.

![cores](https://user-images.githubusercontent.com/6613103/133041217-e8687113-d25c-4526-8b24-980ed5d8a653.png)

The second task is getting far too many cores compared to its itp value. Some of those should have gone to the third task (both itp values are ~ 100, since the third task is running at throughput 200).