@vondele 2021-09-13
@vdbergh now I follow the discussion, and realize the itp algorithm you talk about is mine ;-)

The second task gets much more resources because not all workers can take all tasks. Certain workers (noob's fleet) can right now only do STC tasks because of memory constraints, other workers (more than 32 threads) can only take LTC tasks. So while the itp is providing the order, the task assigned might be different from the top-most one.