@vdbergh 2021-09-13
@vondele Ah sorry. Throughput was introduced by Joona. But I did not know that the current algorithm is yours. Very nice.

I thought about your explanation, but I do not really see why the 2d task would get more cores than the 3d (they are both 10+0.1 so the same machines qualify). For SPSA there is a core limit but as I said I do not think it kicks in.