@vondele 2021-09-13
non-sprt tasks get a 0.5 reduction of their itp, and LLR positive tasks an additional small boost. Not sure if that fully explains it?