@vdbergh 2021-09-13
@vondele 

Yes but that is already incorporated in the itp values. The 2d and 3d task both had itp value ~ 100 (the SPSA job is running at throughput 200).