@vdbergh 2021-09-13
I also checked priorities (both 0).

I was thinking of the following scenario: the 2d task had a high llr at one point and grabbed a bunch of servers from the 3d task (so the effect is doubled). Then its llr came down (and the itp value with it) so that its core count became too high with respect to its itp value. If that's the case then the algorithm is working correctly (after < 15 min the situation should clear).