@ppigazzini 2021-09-13
Just added a SPRT with time odds on DEV

<html>
<body>
<!--StartFragment--><h3 style="margin: 10px 0px; font-family: &quot;Helvetica Neue&quot;, Helvetica, Arial, sans-serif; font-weight: bold; line-height: 40px; color: rgb(177, 173, 167); text-rendering: optimizelegibility; font-size: 24.5px; font-style: normal; font-variant-ligatures: normal; font-variant-caps: normal; letter-spacing: normal; orphans: 2; text-align: start; text-indent: 0px; text-transform: none; white-space: normal; widows: 2; word-spacing: 0px; -webkit-text-stroke-width: 0px; background-color: rgb(17, 17, 17); text-decoration-thickness: initial; text-decoration-style: initial; text-decoration-color: initial;">Active - 3 tests</h3>

21-09-13 | use | master | diff | LLR: 0.05 (-2.94,2.94) <-0.50,2.50>Total: 1488 W: 476 L: 469 D: 543Ptnml(0-2): 54, 167, 290, 184, 49 | sprt @ 0.5+0.005 th 1cores: 26 | #1027 SPRT - priority 5 - throughput 100%
-- | -- | -- | -- | -- | -- | --
21-09-13 | use | master | diff | Pending... (0.2 hrs)<-0.50,2.50> | sprt @ 0.5+0.005 th 1cores: 2 | #1027 SPRT - priority 5 - throughput 100% - time odds 10%
21-09-13 | use | moves1 | diff | 684/30000 iterations1368/60000 games played | 60000 @ 0.5+0.005 th 1cores: 24 | #1027 SPSA - priority 5 - throughput 200%

<!--EndFragment-->
</body>
</html>