@vondele 2021-09-13
yes, the second scenario can also happen, and indeed should clear up after 15min. (unless something stops this process, like reassigning tasks to the last run task automatically).