@vdbergh 2021-09-13
> Just added a SPRT with time odds on DEV
> 
> ### Active - 3 tests
> 21-09-13	use	master	diff	Pending... (0.2 hrs)<-0.50,2.50>	sprt @ 0.5+0.005 th 1cores: 2	#1027 SPRT - priority 5 - throughput 100% - time odds 0.5+0.005 vs 0.55+0.005
> 21-09-13	use	master	diff	LLR: -0.08 (-2.94,2.94) <-0.50,2.50>Total: 21104 W: 6661 L: 6625 D: 7818Ptnml(0-2): 695, 2387, 4376, 2375, 719	sprt @ 0.5+0.005 th 1cores: 26	#1027 SPRT - priority 5 - throughput 100%
> 21-09-13	use	moves1	diff	8152/30000 iterations16304/60000 games played	60000 @ 0.5+0.005 th 1cores: 22	#1027 SPSA - priority 5 - throughput 200%

@ppigazzini I just looked on DEV and currently the 3 runs have similar itp values and similar numbers of cores. So things seems to be correct...