@ppigazzini 2021-09-13
The time odds one (with higher abs(LLR)), always with few cores added:

<html>
<body>
<!--StartFragment--><h3 style="margin: 10px 0px; font-family: &quot;Helvetica Neue&quot;, Helvetica, Arial, sans-serif; font-weight: bold; line-height: 40px; color: rgb(177, 173, 167); text-rendering: optimizelegibility; font-size: 24.5px; font-style: normal; font-variant-ligatures: normal; font-variant-caps: normal; letter-spacing: normal; orphans: 2; text-align: start; text-indent: 0px; text-transform: none; white-space: normal; widows: 2; word-spacing: 0px; -webkit-text-stroke-width: 0px; background-color: rgb(17, 17, 17); text-decoration-thickness: initial; text-decoration-style: initial; text-decoration-color: initial;">Active - 3 tests</h3>

21-09-13 | use | master | diff | LLR: 2.53 (-2.94,2.94) <-0.50,2.50>Total: 3528 W: 1193 L: 991 D: 1344Ptnml(0-2): 71, 350, 770, 452, 121 | sprt @ 0.5+0.005 th 1cores: 20 | #1027 SPRT - priority 5 - throughput 100% - time odds 0.5+0.005 vs 0.55+0.005
-- | -- | -- | -- | -- | -- | --
21-09-13 | use | master | diff | LLR: 0.28 (-2.94,2.94) <-0.50,2.50>Total: 24848 W: 7848 L: 7775 D: 9225Ptnml(0-2): 814, 2786, 5178, 2805, 841 | sprt @ 0.5+0.005 th 1cores: 16 | #1027 SPRT - priority 5 - throughput 100%
21-09-13 | use | moves1 | diff | 9380/30000 iterations18760/60000 games played | 60000 @ 0.5+0.005 th 1cores: 14 | #1027 SPSA - priority 5 - throughput 200%

<!--EndFragment-->
</body>
</html>